+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 
 - Upgraded Android SDK to Level 34
+- Upgraded Android Gradle Plugin to 8.1.3
 
 
 ## [v1.2.7] (60) 㨆莊 (2023-11-05)