+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## [Unreleased]
 
 - Upgraded Android SDK to Level 34
-- Upgraded Android Gradle Plugin to 8.1.3
+- Upgraded Android Gradle Plugin to 8.1.4
 - Upgraded Material Components For Android to 1.10.0
 - Fixed `DisplayMetrics.scaledDensity` deprecated in API level 34